tests/wpt/tests/css/css-grid/grid-lanes/tentative/parsing/grid-lanes-pack-computed.html

<!DOCTYPE html>
<html>
<head>
  <meta charset="utf-8">
  <title>CSS Grid Lanes: grid-lanes-pack getComputedStyle()</title>
  <link rel="author" title="Alison Maher" href="mailto:almaher@microsoft.com">
  <link rel="help" href="https://drafts.csswg.org/css-grid-3">
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
  <script src="/css/support/computed-testcommon.js"></script>
  <script src="/css/support/inheritance-testcommon.js"></script>
</head>
<body>
  <div id="target"></div>
  </div>
  <script>
    test_computed_value("grid-lanes-pack", "normal");
    test_computed_value("grid-lanes-pack", "dense");
  </script>
</body>
</html>
